mysql uses long, not number

+++ src/main/resources/db/createDB.sql
@@ -4,6 +4,6 @@ create table trainings (
 	 id varchar(255), 
 	 team_login varchar(255), 
 	 name varchar(255), 
-	 distance number(10), 
-	 training_date number(19)
+	 distance long,
+	 training_date long
 );
